+++ src/dsm-cc/descriptors/descriptors.h
@@ -12,7 +12,7 @@ bool dsmcc_descriptor_is_parseable(struct dentry *dentry, uint8_t tag, int expec
 struct dsmcc_descriptor *dsmcc_descriptors_init(struct demuxfs_data *priv);
 void dsmcc_descriptors_destroy(struct dsmcc_descriptor *descriptor_list);
 struct dsmcc_descriptor *dsmcc_descriptors_find(uint8_t tag, struct demuxfs_data *priv);
-uint8_t dsmcc_descriptors_parse(const char *payload, uint8_t num_descriptors, 
+uint32_t dsmcc_descriptors_parse(const char *payload, uint8_t num_descriptors, 
 		struct dentry *parent, struct demuxfs_data *priv);
 
 /* Descriptor parsers */
