+++ src/tables/descriptors/descriptors.h
@@ -15,7 +15,7 @@ struct descriptor *descriptors_init(struct demuxfs_data *priv);
 void descriptors_destroy(struct descriptor *descriptor_list);
 struct descriptor *descriptors_find(uint8_t tag, struct demuxfs_data *priv);
 int descriptors_count(const char *payload, uint16_t program_information_length);
-uint8_t descriptors_parse(const char *payload, uint8_t num_descriptors, 
+uint32_t descriptors_parse(const char *payload, uint8_t num_descriptors, 
 		struct dentry *parent, struct demuxfs_data *priv);
 
 /* Descriptor parsers */
